#4936 #7152 even better tests
