+++ tests/ZendTest/Mvc/Router/PriorityListTest.php
@@ -11,6 +11,7 @@ namespace ZendTest\Mvc\Router;
 
 use Zend\Mvc\Router\PriorityList;
 use PHPUnit_Framework_TestCase as TestCase;
+use Zend\Stdlib\ArrayUtils;
 
 /**
  * @group      Zend_Router
@@ -136,5 +137,16 @@ class PriorityListTest extends TestCase
         }
 
         $this->assertEquals(array('bar', 'foo', 'baz'), $orders);
+    }
+
+    public function testKey()
+    {
+        $this->list->insert('bar', 'foo', 0);
+        $this->list->insert('baz', 'f00', 2);
+
+        $list = ArrayUtils::iteratorToArray($this->list);
+
+        end($list);
+        $this->assertEquals('bar', key($list));
     }
 }
